etnaviv: fix computation of source_offset[1] in etna_try_rs_blit()
The source offset for the second pixel pipe should be half way down
the height to be copied, not half way through the size of the buffer
level.  The size of the buffer level includes all the depth and array
attributes, so the old calculation means that the second pixel pipe
will be several layers of depth below where it should be.

Fix this.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>