etnaviv: eliminate the second destination BO
Eliminate passing a second destination BO to etna_compile_rs_state().
This should not be necessary, as the second BO is used when the BO
is multi-tiled, and we're clearly only talking about a single BO.
The destination offset for the second half of the tiled surface can
also be computed by etna_compile_rs_state(), so reduce the dest_offset
array to a single plain integer to avoid any future mistakes.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>