etnaviv: fix vertex Z position coordinate fixup
The vertex Z position coordinate is fixed up on older GPUs by the
following:

	Z' = (Z + 1) / 2

Rather than using a '1' uniform, we use the W position component.
However, the W position component is not guaranteed to always be '1' -
the preceding vertex shader code can modify this. Rather than using the
W component as a short-cut to gain a '1' constant, use an explicit '1'
uniform instead.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>