etnaviv: fix FLR implementation for older GPUs, make CEIL and SIGN bug
Older GPUs do not support the FLR instruction, but we still generate
its opcode in the shader instruction stream. Add a FLR implementation
for these CPUs using  using FRC and ADD for older GPUs. This is
implemented as:

	FRC tX.xyzw, void, void tSRC.????
	ADD tDST.????, tSRC.????, void, -tX.xyzw

Other instructions (CEIL and SIGN) also need to be fixed, but currently
we make them behave as other unimplemented instructions.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>