etnaviv: locally cache the ctx->stream pointer in etna_emit_state()
In order for EMIT_STATE*() macros to be usable in etna_submit_rs_state,
we need to have this and etna_emit_state() behaving the same way.
Change etna_emit_state() to cache the stream pointer in a local
variable in the same way that etna_submit_rs_state() is.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>