etnaviv: another fix for width/height in etna_try_rs_blit()
We should not be using the source width/height when copying:
* what if the destination is smaller than the source?
* what if the requested box is smaller than the source?

The result can be memory corruption, or corruption of the destination
resource.  This causes amoeba to fail when supertiling enabled, and we
are trying to convert between different tiling formats (with their
different alignments.)

Instead, use the requested source box size, but if the width is not
aligned to the RS width and is within the padding region, increase it
to fit the RS restrictions.  This allows us to work on both GC2000 with
supertiling enabled, and GC600 without overflowing the source or
destination buffer sizes.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>