etnaviv: convert sampled textures to correct tiling format
Multi-pipe GPUs can only render to multi-tiled or super-multi-tiled
resources, but the samplers can only read from tiled formats. If the
resource we are given to sample is not in the correct tile format,
allocate a new resource, and when binding, appropriately de-tile the
original resource so the GPU can read it.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>