etnaviv: avoid rendering corruption and hangs with amoeba on GC2000
Avoid rendering corruption with amoeba on the GC2000 noticed during
the opening rendering of PI.  Amoeba renders PI from right to left,
beginning with digits 8, 8, 2.  When the second and third digits are
rendered, the first and second (respectively) are corrupted with noise.

Further problems occur after PI has finished rendering, and we switch
to rendering the "title" screens, with corruption and recoverable GPU
hangs.

Adding cache flushes and RA/PE semaphore/stalls do not appear to have
any effect on the problem, but enabling draw_stall fixes it.  Further
experimentation shows that the problem occurs when two draw operations
follow each other, even when they are very similar draws.  Analysis
of the second draw shows that it touches:
 - vertex element configuration
 - index address & type
 - vertex address & stride
 - VS config
 - PA shader attributes
 - PS config
 - varying config
 - VS instructions & uniforms
 - PS instructions (but no PS uniforms)

Experimentation shows that adding a FE/PE stall prior to the draw OP
resolves all these problems.  Further experimentation with the location
of the stall shows that the problem states are the shader instructions
and maybe uniforms too.

This does not appear to affect GC600.

Given this, key the additional stall off loading the Vertex code at
0xc000 (the new shader layout) and reloading the shader or uniforms.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>