etnaviv: allow transfer map to access texture surface
There are two paths used to update textures: one of them is by using
the GPU, the other is via transfer_map().  We want to avoid bouncing
updates via the render buffer when using transfer_map(), and we have
to cater for the case where we want to render into a texture which
has been updated via transfer_map().

In order for this to work, we need to detect when the texture is
younger than the render buffer, and bring the render buffer up to
date.  Add a hook to update the render buffer from the texture
buffer.

Add hooks to transfer_map() and transfer_unmap() to update the texture
buffer instead of the render buffer if it exists, and is the same
age or younger than the render buffer.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>