u_vbuf: don't try to set unsupported vertex buffers
If hardware supports a limited number of vertex buffers, we should not
try to pass all buffers to the gallium driver then we are in buffer
translation mode.

This can happen when there are multiple vertex buffers, but the
hardware only supports one vertex buffer (eg, GC600.)

Mask the set of dirty buffers with the hardware supported buffer mask
before determining the start/count of vertex buffers to be updated.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Reviewed-by: Christian Gmeiner <christian.gmeiner@gmail.com>