etnaviv: emit correct number of vertex buffers
The number of vertex buffers that a Vivante GPU supports is defined by
the stream count, not by the presence of the shader range registers.
Arrange to only write the actively in-use vertex buffer registers,
skipping the register at 0x680/0x6a0 as this does not appear to be
used.

Tested on GC600, GC880, and GC2000.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>