etnaviv: disable flatshading on GC880+
It seems that GC880+ has a hardware bug, which disables all
varying interpolation if flatshading is enabled. I guess that
this has not really been noticed as GL_SMOOTH is the default
shading mode and GLES >= 2.0 does not allow to change it anymore.

The only thing that is requesting flatshading (aside from old
apps and tests) is the gallium u_blitter, breaking stuff like
mipmap generation.

Fix this by ignoring the mostly obsolete flatshading on those
cores.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>