etnaviv: fix shader program maximum inputs
We currently assume that all shaders on all devices can accept 16
inputs. However, the number of inputs for a shader program is
limited by the data source: for a vertex shader, it is the maximum
number of vertex elements, and for a fragment shader, it is the
maximum number of varyings.

Hence, for older GPUs, this is limited to 10 inputs for the vertex
shader and 8 inputs for the fragment shader.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Wladimir J. van der Laan <laanwj@gmail.com>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>