etnaviv: fix MSAA blit dimensions
The MSAA blitting is buggy because it assumes that the box dimensions
are in samples.  This is untrue: the box dimensions are in pixels,
where a pixel may have more than one sample.

Since we do not support scaling or flipping, the source box size must
match the destination box size, so reject any blit with differing box
sizes.

As the hardware takes the number of source samples to be blitted, and
we only support downsampling, use the source box to determine the size
of the blit, and validate this against the source and destination
resources.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>