etnaviv: correct max render and max texture sizes
For GPUs without the 8K texture and 8K render capabilities, the maximum
size is 2048 not 4096.  Providing such GPUs with a 2048x2048 texture
results in rendering errors, as found via the piglit fbo-maxsize test.
Reducing this to 2048x2048 results in it working correctly.

Vivante's library also reports that the max texture and max render
sizes for GC600 are both 2048x2048.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>