etnaviv: fix samplers and sampler views properly this time
Nothing says that sampler 0 has to be the first one to be used; mesa
can provide a configuration for sampler 1 without a configuration for
sampler 0.  This leads to a segfault when we try to dereference the
resulting NULL sampler and sampler view.

While looking at this, there's other bugs too here: we do not
handle the vertex shader textures properly, neither do we properly
set the number of texture samplers for the vertex shader.

Switch over to using a bitmask of the active samplers and sampler
views, which then simplifies active_samplers_bits().

We need to keep num_fragment_sampler_views and num_fragment_samplers
for the u_blitter code.

This helps the piglit max-samplers test to pass.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Acked-by: Christian Gmeiner <christian.gmeiner@gmail.com>