etnaviv: bring back support for const buffer
In commit d41cc12 the const buffer support got broken to be able to do
some refactorings in the area of uniform handling. With this commit full
const buffer support is back. Const buffer updates are handled with a
seperate dirty flag, are seperated from shader emit code and the
initial state of all const buffer values is 0.

Fixes piglit tests using glUniform*.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>