etnaviv: keep track of dirty flags affecting uniforms
To support e.g. rect textures we need to react on driver state
changes affecting uniforms. That could be CONSTBUF or SHADER but in case
of rect textures we need to reload unifroms if SAMPLER states have
changed. This commit adds etna_set_shader_uniforms_dirty_flags(..)
to determine which dirty flags triggers an uniforms reload.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>