etnaviv: improve etna_rasterizer_state_create(..)
We do not support all features which are defined by pipe_rasterizer_state.
If such a case happens we do simply return NULL and the application will
crash as the state tracker does not validate the return value. As a
crashing application is all the worse then a miss-rendering one we return
a valid state object.

Fixes gl-1.0-polygon-line-aa piglit.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>